Cleanup for size_t, int64_t
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@14440 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/src/c++/client/protocol/socket/SocketDriver.cpp
@@ -251,7 +251,7 @@ SocketDriver::SocketDriver(Global& global, const string instanceName)
 
    global_.fillArgs(*argsStructP_);
    try {
-      connection_ = getXmlBlasterAccessUnparsed(argsStructP_->argc, argsStructP_->argv);
+      connection_ = getXmlBlasterAccessUnparsed((int)argsStructP_->argc, argsStructP_->argv);
       if (connection_) {
          connection_->userObject = this; // Transports us to the myUpdate() method
          connection_->log = myLogger;    // Register our own logging function
@@ -289,7 +289,7 @@ void SocketDriver::reconnectOnIpLevel(void)
    ::ExceptionStruct socketException;
 
    try {
-      connection_ = getXmlBlasterAccessUnparsed(argsStructP_->argc, argsStructP_->argv);
+      connection_ = getXmlBlasterAccessUnparsed((int)argsStructP_->argc, argsStructP_->argv);
       connection_->userObject = this; // Transports us to the myUpdate() method
       connection_->log = myLogger;    // Register our own logging function
       connection_->logUserP = this;   // Pass ourself to myLogger()
